Admin part is done

+++ src/main/java/eng/hospitalSystemService/web/servlets/DeleteDepartmentServlet.java
@@ -1,8 +1,6 @@
 package eng.hospitalSystemService.web.servlets;
 
 import eng.hospitalSystemService.app.*;
-import eng.hospitalSystemService.db.entities.PersonalEntity;
-import eng.hospitalSystemService.db.entities.PokojEntity;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -10,7 +8,6 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
-import java.util.List;
 
 @WebServlet(name = "DeleteDepartmentServlet",urlPatterns = "/deleteDepartment")
 public class DeleteDepartmentServlet extends HttpServlet {
@@ -27,35 +24,19 @@ public class DeleteDepartmentServlet extends HttpServlet {
             response.sendRedirect("listOfDepartments.jsp");
         }
         else {
-            try {
-                PersonalService personalService = new PersonalService();
-                List<PersonalEntity> personalEntityList = personalService.getListOfPersonal();
-                for (PersonalEntity personalEntity: personalEntityList) {
-                    if (personalEntity.getDepartment()==departmentID){
-                        alertService.add(Alert.Type.danger, "There are assigned Personals please reassign them before deleting department.");
-                    }
+            if (departmentService.isDepartmentEmpty(departmentID))
+            {
+                try {
+                    departmentService.delete(departmentID);
+                } catch (Exception e) {
+                    throw new RuntimeException("Failed to delete department via Servlet.", e);
                 }
+                alertService.add(Alert.Type.success,"department was deleted.");
+                response.sendRedirect("listOfDepartments.jsp");
             }
-            catch (Exception e){
-                throw new RuntimeException("Failed to delete due Personal in this department.",e);
-            }
-
-            try {
-                RoomService roomService = new RoomService();
-                List<PokojEntity> rooms = roomService.getAllRooms();
-                for (PokojEntity pokoj: rooms) {
-                    if (pokoj.getDepartmentid()==departmentID)
-                        alertService.add(Alert.Type.danger, "There are assigned Rooms please reassign them before deleting department.");
-                }
-            }catch (Exception e){throw new RuntimeException("Failed to delete due rooms in this department.");}
-
-            try {
-                departmentService.delete(departmentID);
-            } catch (Exception e) {
-                throw new RuntimeException("Failed to delete department via Servlet.", e);
-            }
+            else{alertService.add(Alert.Type.danger,"Department is not empty. Please reassign personal or rooms from it.");}
         }
-        response.sendRedirect("deletedDep.jsp");
+        response.sendRedirect("departmentInventory.jsp?departmentID="+departmentID);
     }
 
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
